+++ src/tquality_core/assets/js_scripts/element/blur.js
@@ -1,0 +1,9 @@
+/* global arguments */
+
+/** @type {HTMLElement} */
+const element = arguments[0];
+
+element.dispatchEvent(new Event('blur', {bubbles: true}));
+if (document.activeElement instanceof HTMLElement) {
+    document.activeElement.blur();
+}